backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.1.1
whitenoise==6.7.0
gunicorn==23.0.0
